Python_Churn_prediction/ChurnPrediction/requirements.txt
mysql-connector-python
sqlalchemy
pymysql
scikit-learn
matplotlib
seaborn